% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed.R
\name{get_parameters}
\alias{get_parameters}
\title{Get item parameters}
\usage{
get_parameters(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row is an item
}
\description{
\code{get_parameters()} returns information about the item IRT parameters that were used for creating scores. See the \href{https://levante-framework.github.io/rlevante/index.html}{rlevante documentation} for more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
parameters <- get_parameters(data_source = "levante_data_example:d0rt", version = "current")
}
}
