% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed.R
\name{get_participants}
\alias{get_participants}
\title{Get participants}
\usage{
get_participants(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row contains information about a child
participant.
}
\description{
\code{get_participants()} returns information about participants as a data frame.
See the \href{https://levante-framework.github.io/levante-r/index.html}{levante-r documentation} for
more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
participants <- get_participants(data_source = "levante_data_example:d0rt", version = "current")
}
}
