% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed.R
\name{get_scores}
\alias{get_scores}
\title{Get scores}
\usage{
get_scores(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row is a task ability score. See our
\href{https://researcher.levante-network.org/measures/scoring-and-psychometrics}{Scoring and Psychometrics page}
to learn how to interpret scores.
}
\description{
\code{get_scores()} returns information about scores as a data frame. See the
\href{https://levante-framework.github.io/levante-r/index.html}{levante-r documentation} for
more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
scores <- get_scores(data_source = "levante_data_example:d0rt", version = "current")
}
}
