% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-processed.R
\name{get_trials}
\alias{get_trials}
\title{Get trials}
\usage{
get_trials(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row is a trial.
}
\description{
\code{get_trials()} returns information about each trial as a data frame. See the
\href{https://levante-framework.github.io/levante-r/index.html}{levante-r documentation} for
more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
trials <- get_trials(data_source = "levante_data_example:d0rt", version = "current")
}
}
