% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{get_variables}
\alias{get_variables}
\title{Get variable metadata}
\usage{
get_variables(data_source, version = "current")
}
\arguments{
\item{data_source}{Name of Redivis dataset}

\item{version}{Version of Redivis dataset}
}
\value{
A data frame where each row is a variable.
}
\description{
\code{get_variables()} returns metadata about the variables in a Redivis dataset.
See the \href{https://levante-framework.github.io/levante-r/index.html}{levante-r documentation} for
more information about how to access LEVANTE datasets and codebooks.
}
\examples{
\dontrun{
variables <- get_variables(data_source = "levante_data_example:d0rt", version = "current")
}
}
