% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levante-package.R
\docType{package}
\name{levante-package}
\alias{levante}
\alias{levante-package}
\title{levante: Data access for the LEVANTE data repository}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

}
\keyword{internal}
